warmup lr

+++ models/dynamic_rnn.py
@@ -185,8 +185,10 @@ class DRNN(object):
             else:
                 self.loss = self.xent_loss
             self.warmup_step = 4000
-            self.learning_rate = tf.sqrt(config.model_size) * tf.minimum(
-                1 / tf.sqrt(self.global_step), tf.pow(self.warmup_step, -1.5))
+            self.learning_rate = tf.sqrt(
+                tf.cast(config.model_size, tf.float32)) * tf.minimum(
+                1 / tf.sqrt(tf.cast(self.global_step, tf.float32)),
+                tf.pow(self.warmup_step, -1.5))
             self.optimizer = tf.train.AdamOptimizer(self.learning_rate, 0.9,
                                                     0.98)
